% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_format_time_minutes}
\alias{fads_format_time_minutes}
\title{FADS Format Time Into Minutes}
\usage{
fads_format_time_minutes(fads)
}
\arguments{
\item{fads}{TODO}
}
\value{
TODO
}
\description{
FADS Format Time Into Minutes
}
