% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_format_time_seconds}
\alias{fads_format_time_seconds}
\title{FADS Format Time Into Seconds}
\usage{
fads_format_time_seconds(fads)
}
\arguments{
\item{fads}{TODO}
}
\value{
TODO
}
\description{
FADS Format Time Into Seconds
}
