% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_generate_report}
\alias{fads_generate_report}
\title{Generate Report from a FADS TSV File}
\usage{
fads_generate_report(tsvFile, bins = 100)
}
\arguments{
\item{tsvFile}{FADS TSV file.}

\item{bins}{Number of bins for the density tile size.}
}
\value{
A patchwork object
}
\description{
TODO make sure that all necessary functions imported
}
