% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_get_outliers}
\alias{fads_get_outliers}
\title{FADS Get Outliers}
\usage{
fads_get_outliers(numbers, scale = 5)
}
\arguments{
\item{numbers}{TODO}

\item{scale}{Scale factor that is used to calculate how much a datapoint can
deviate from the bulk before it is considered an outlier.}
}
\value{
TODO
}
\description{
Returns TRUE for those elements that are less than Q1 - scale * IQR or more
than Q3 + scale * IQR, and FALSE otherwise.
}
