% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_marker_opacity}
\alias{fads_marker_opacity}
\title{FADS Get Marker Opacity}
\usage{
fads_marker_opacity(fads)
}
\arguments{
\item{fads}{TODO}
}
\value{
An opacity value between 0 and 1.
}
\description{
The purpose of this function is to determine a suitable alpha value between 0
and 1 so that scatter plots with a large number of points become meaningful.
}
