% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_plot_frequency}
\alias{fads_plot_frequency}
\title{FADS Plot Droplet Frequency}
\usage{
fads_plot_frequency(fads)
}
\arguments{
\item{fads}{Tibble as returned from \code{fads_read_tsv()}.}
}
\value{
A ggplot object
}
\description{
FADS Plot Droplet Frequency
}
