% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_plot_red_green_hex}
\alias{fads_plot_red_green_hex}
\title{FADS Plot Red vs. Green Max PMT}
\usage{
fads_plot_red_green_hex(fads, bins = 100)
}
\arguments{
\item{fads}{Tibble as returned from \code{fads_read_tsv()}.}

\item{bins}{Argument passed on to \code{ggplot2::geom_bin2d()}.}
}
\value{
A ggplot object.
}
\description{
TODO plot elapsed time
}
\details{
This is a copy of \code{fads_plot_red_green()}, just with hexagonal tiles.
}
