% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_plot_width}
\alias{fads_plot_width}
\title{FADS Plot With against Red/Green Signal}
\usage{
fads_plot_width(fads, y = "red", bins = 100)
}
\arguments{
\item{fads}{Tibble as returned from \code{fads_read_tsv()}.}

\item{y}{Name of variable to plot on the y-axis.}

\item{bins}{Argument passed on to \code{ggplot2::geom_bin2d()}.}
}
\value{
A ggplot object.
}
\description{
TODO plot elapsed time
}
