% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_plot_width_time}
\alias{fads_plot_width_time}
\title{FADS Plot Width vs. Time}
\usage{
fads_plot_width_time(fads)
}
\arguments{
\item{fads}{Tibble as returned from \code{fads_read_tsv()}.}
}
\value{
A ggplot object.
}
\description{
FADS Plot Width vs. Time
}
