% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_read_all}
\alias{fads_read_all}
\title{FADS Read All}
\usage{
fads_read_all(dir, suffix)
}
\arguments{
\item{dir}{Directory containing FADS TSV files.}

\item{suffix}{Suffix identifiying FADS TSV files.}
}
\value{
A tibble.
}
\description{
Combines all available data into one tibble.
}
