% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_read_tsv}
\alias{fads_read_tsv}
\title{FADS Read TSV File}
\usage{
fads_read_tsv(fadsData)
}
\arguments{
\item{fadsData}{Path to FADS CSV file.}
}
\value{
Tibble containing data about droplets identified by the sorter.
}
\description{
FADS Read TSV File
}
