% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_remove_outliers}
\alias{fads_remove_outliers}
\title{FADS Remove Outliers}
\usage{
fads_remove_outliers(fads)
}
\arguments{
\item{fads}{TODO.}
}
\value{
TODO
}
\description{
Removes rows that contain green or red outliers.
}
