% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fads.R
\name{fads_reports_for_all}
\alias{fads_reports_for_all}
\title{Generate a FADS Report PDF For Every File in Folder}
\usage{
fads_reports_for_all(directory, outputDir)
}
\arguments{
\item{directory}{A directory.}

\item{outputDir}{Destination for the generated PDF files.}
}
\description{
This function exploits that the first line in every FADS TSV file is:
# iiiiiiiiiiiiii0000000
}
