% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{logfile_align_pressure}
\alias{logfile_align_pressure}
\title{Align Pressure to Syringe Log}
\usage{
logfile_align_pressure(pressureLog, syringeLog, tPressure, tSyringe)
}
\arguments{
\item{pressureLog}{TODO}

\item{syringeLog}{TODO}

\item{tPressure}{In seconds.}

\item{tSyringe}{In seconds.}
}
\value{
TODO
}
\description{
Align Pressure to Syringe Log
}
