% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{logfile_plot_pressure}
\alias{logfile_plot_pressure}
\title{Plot Pressure Log}
\usage{
logfile_plot_pressure(pressureLog)
}
\arguments{
\item{pressureLog}{Pressure log.}
}
\value{
A ggplot object.
}
\description{
Plot Pressure Log
}
