% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{logfile_plot_syringepump}
\alias{logfile_plot_syringepump}
\title{Plot Syringe Pump Log}
\usage{
logfile_plot_syringepump(syringepumpLog)
}
\arguments{
\item{syringepumpLog}{Syringe pump log.}
}
\value{
A ggplot object.
}
\description{
Plot Syringe Pump Log
}
