% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{logfile_read_pressure}
\alias{logfile_read_pressure}
\title{Read Elveflow OB1 Pressure Controller Log File}
\usage{
logfile_read_pressure(filename)
}
\arguments{
\item{filename}{Path to the log file.}
}
\value{
Channel pressures in long format.
}
\description{
Read Elveflow OB1 Pressure Controller Log File
}
