% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{logfile_read_syringepump}
\alias{logfile_read_syringepump}
\title{Read Cetoni Syringe Pump Log File}
\usage{
logfile_read_syringepump(filename)
}
\arguments{
\item{filename}{Path to the log file.}
}
\value{
Syringe flowrates in long format.
}
\description{
Read Cetoni Syringe Pump Log File
}
