% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spincurves.R
\name{plot_heightprofile}
\alias{plot_heightprofile}
\title{Plot Height Profile}
\usage{
plot_heightprofile(heights)
}
\arguments{
\item{heights}{Dataframe containing columnes 'position' and 'height'}
}
\value{
A ggplot object
}
\description{
Plot Height Profile
}
