% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spincurves.R
\name{plot_heightprofile_bywafer}
\alias{plot_heightprofile_bywafer}
\title{Plot Height Profiles By Wafer}
\usage{
plot_heightprofile_bywafer(heights)
}
\arguments{
\item{heights}{TODO}
}
\value{
A ggplot object.
}
\description{
Plot Height Profiles By Wafer
}
