% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_pretty_custom}
\alias{theme_pretty_custom}
\title{Custom ggplot2 Theme For Package Figures}
\usage{
theme_pretty_custom()
}
\value{
A patchwork object.
}
\description{
Custom ggplot2 Theme For Package Figures
}
